+++ airbyte-cdk/python/airbyte_cdk/sources/singer/singer_helpers.py
@@ -124,7 +124,9 @@ class SingerHelper:
             field_object["type"] = SingerHelper._parse_type(field_object["type"])
 
     @staticmethod
-    def singer_catalog_to_airbyte_catalog(singer_catalog: Dict[str, any], sync_mode_overrides: Dict[str, SyncModeInfo], primary_key_overrides: Dict[str, List[str]]) -> AirbyteCatalog:
+    def singer_catalog_to_airbyte_catalog(
+        singer_catalog: Dict[str, any], sync_mode_overrides: Dict[str, SyncModeInfo], primary_key_overrides: Dict[str, List[str]]
+    ) -> AirbyteCatalog:
         """
         :param singer_catalog:
         :param sync_mode_overrides: A dict from stream name to the sync modes it should use. Each stream in this dict must exist in the Singer catalog,
@@ -139,34 +141,41 @@ class SingerHelper:
             airbyte_stream = AirbyteStream(name=name, json_schema=schema)
             if name in sync_mode_overrides:
                 override_sync_modes(airbyte_stream, sync_mode_overrides[name])
-
             else:
                 set_sync_modes_from_metadata(airbyte_stream, stream.get("metadata", []))
 
             if name in primary_key_overrides:
                 airbyte_stream.source_defined_primary_key = [[k] for k in primary_key_overrides[name]]
-            else:
-                if stream.get("key_properties"):
-                    airbyte_stream.source_defined_primary_key = [[k] for k in stream["key_properties"]]
+            elif stream.get("key_properties"):
+                airbyte_stream.source_defined_primary_key = [[k] for k in stream["key_properties"]]
 
             airbyte_streams += [airbyte_stream]
         return AirbyteCatalog(streams=airbyte_streams)
 
     @staticmethod
-    def get_catalogs(logger, shell_command: str, sync_mode_overrides: Dict[str, SyncModeInfo], primary_key_overrides: Dict[str, List[str]], excluded_streams: List) -> Catalogs:
+    def _read_singer_catalog(logger, shell_command: str) -> Mapping[str, Any]:
         completed_process = subprocess.run(
             shell_command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE, universal_newlines=True
         )
-
         for line in completed_process.stderr.splitlines():
             logger.log_by_prefix(line, "ERROR")
 
-        singer_catalog = json.loads(completed_process.stdout)
+        return json.loads(completed_process.stdout)
+
+    @staticmethod
+    def get_catalogs(
+        logger,
+        shell_command: str,
+        sync_mode_overrides: Dict[str, SyncModeInfo],
+        primary_key_overrides: Dict[str, List[str]],
+        excluded_streams: List,
+    ) -> Catalogs:
+        singer_catalog = SingerHelper._read_singer_catalog(logger, shell_command)
         streams = singer_catalog.get("streams", [])
         if streams and excluded_streams:
             singer_catalog["streams"] = [stream for stream in streams if stream["stream"] not in excluded_streams]
-        airbyte_catalog = SingerHelper.singer_catalog_to_airbyte_catalog(singer_catalog, sync_mode_overrides, primary_key_overrides)
 
+        airbyte_catalog = SingerHelper.singer_catalog_to_airbyte_catalog(singer_catalog, sync_mode_overrides, primary_key_overrides)
         return Catalogs(singer_catalog=singer_catalog, airbyte_catalog=airbyte_catalog)
 
     @staticmethod