+++ airbyte-cdk/python/unit_tests/singer/test_singer_helpers.py
@@ -23,111 +23,54 @@
 #
 
 
+import copy
+
 from airbyte_cdk.sources.singer import SingerHelper
 
-
-def test_singer_catalog_to_airbyte_catalog():
-    singer_catalog = {
-        "streams": [
-            {
+basic_singer_catalog = {
+    "streams": [
+        {
             "type": "SCHEMA",
             "stream": "users",
             "schema": {
                 "properties": {
-                "id": {
-                    "type": "integer"
-                },
-                "name": {
-                    "type": "string"
-                },
-                "updated_at": {
-                    "type": "string",
-                    "format": "date-time"
-                }
+                    "id": {"type": "integer"},
+                    "name": {"type": "string"},
+                    "updated_at": {"type": "string", "format": "date-time"},
                 }
             },
             "key_properties": ["id"],
-            "bookmark_properties": ["updated_at"]
-            }
-        ]
-    }
+            "bookmark_properties": ["updated_at"],
+        }
+    ]
+}
+
+
+def test_singer_catalog_to_airbyte_catalog():
     airbyte_catalog = SingerHelper.singer_catalog_to_airbyte_catalog(
-        singer_catalog=singer_catalog,
-        sync_mode_overrides={},
-        primary_key_overrides={}
+        singer_catalog=basic_singer_catalog, sync_mode_overrides={}, primary_key_overrides={}
     )
 
     user_stream = airbyte_catalog.streams[0]
-    assert user_stream.source_defined_primary_key == [['id']]
+    assert user_stream.source_defined_primary_key == [["id"]]
 
 
-def test_singer_catalog_to_airbyte_catalog_composite_primary_key():
-    singer_catalog = {
-        "streams": [
-            {
-            "type": "SCHEMA",
-            "stream": "users",
-            "schema": {
-                "properties": {
-                "id": {
-                    "type": "integer"
-                },
-                "name": {
-                    "type": "string"
-                },
-                "updated_at": {
-                    "type": "string",
-                    "format": "date-time"
-                }
-                }
-            },
-            "key_properties": ["id", "name"],
-            "bookmark_properties": ["updated_at"]
-            }
-        ]
-    }
+def test_singer_catalog_to_airbyte_catalog_composite_pk():
+    singer_catalog = copy.deepcopy(basic_singer_catalog)
+    singer_catalog["streams"][0]["key_properties"] = ["id", "name"]
 
     airbyte_catalog = SingerHelper.singer_catalog_to_airbyte_catalog(
-        singer_catalog=singer_catalog,
-        sync_mode_overrides={},
-        primary_key_overrides={}
+        singer_catalog=singer_catalog, sync_mode_overrides={}, primary_key_overrides={}
     )
 
     user_stream = airbyte_catalog.streams[0]
-    assert user_stream.source_defined_primary_key == [['id'], ['name']]
+    assert user_stream.source_defined_primary_key == [["id"], ["name"]]
 
 
-def test_singer_catalog_to_airbyte_catalog_primary_key_override():
-    singer_catalog = {
-        "streams": [
-            {
-            "type": "SCHEMA",
-            "stream": "users",
-            "schema": {
-                "properties": {
-                "id": {
-                    "type": "integer"
-                },
-                "name": {
-                    "type": "string"
-                },
-                "updated_at": {
-                    "type": "string",
-                    "format": "date-time"
-                }
-                }
-            },
-            "key_properties": ["id", "name"],
-            "bookmark_properties": ["updated_at"]
-            }
-        ]
-    }
-
+def test_singer_catalog_to_airbyte_catalog_pk_override():
     airbyte_catalog = SingerHelper.singer_catalog_to_airbyte_catalog(
-        singer_catalog=singer_catalog,
-        sync_mode_overrides={},
-        primary_key_overrides={'users': ['name']}
+        singer_catalog=basic_singer_catalog, sync_mode_overrides={}, primary_key_overrides={"users": ["name"]}
     )
 
     user_stream = airbyte_catalog.streams[0]
-    assert user_stream.source_defined_primary_key == [['name']]
+    assert user_stream.source_defined_primary_key == [["name"]]