+++ airbyte-cdk/python/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
+## 0.1.7
+Allow to fetch primary key info from singer catalog
+
 ## 0.1.6
-Add abstraction for creating destinations. 
+Add abstraction for creating destinations.
 
 Fix logging of the initial state.